Don't try to monitor fast path links unless configured.

+++ projectw.sh
@@ -328,21 +328,25 @@ if [ "$ACTION" != "RESET" ]; then
       fi
     done
 
-    # Add any fastpath links to dp0 and verify that the get the port number we expect
-    tail -n +2 $DP0LINKS | while IFS=, read interface dp0_port
-    do
-        # Match the fastpath connection to the directly connected ports upon that switch
-        # Add the interface
-        $VSCTL add-port $RFDP $interface -- set Interface $interface ofport_request="$dp0_port"
-        # Verify that it got the correct port number
-        new_port=$(get_dp0_port $interface)
-        if [ "$new_port" != "$dp0_port" ]; then
-            echo_bold "Error: fastpath mapped to wrong OF port $interface required =$dp0_port but instead is =$new_port"
-            kill -INT $$
-        else
-            echo_bold "Successfully added $interface as port $new_port"
-        fi
-    done
+    if [ -f $DP0LINKS ] ; then 
+        # Add any fastpath links to dp0 and verify that the get the port number we expect
+        tail -n +2 $DP0LINKS | while IFS=, read interface dp0_port
+        do
+            # Match the fastpath connection to the directly connected ports upon that switch
+            # Add the interface
+            $VSCTL add-port $RFDP $interface -- set Interface $interface ofport_request="$dp0_port"
+            # Verify that it got the correct port number
+            new_port=$(get_dp0_port $interface)
+            if [ "$new_port" != "$dp0_port" ]; then
+                echo_bold "Error: fastpath mapped to wrong OF port $interface required =$dp0_port but instead is =$new_port"
+                kill -INT $$
+            else
+                echo_bold "Successfully added $interface as port $new_port"
+            fi
+        done
+    else
+        echo_bold "NOT using fast path (will use packet in)"
+    fi
 
     # Setup inband control channel
     if [ -n "$ICC_INT" ]; then
